.zshrc updates
- shorten shell prompt
- conda init followed by conda deactivate
- add ~/dev/bin to PATH

- typeset -U path PATH

https://zsh.sourceforge.io/Guide/zshguide02.html\#l24

The incantation `typeset -U path', where the -U stands for unique,
tells the shell that it should not add anything to $path if it's
there already. To be precise, it keeps only the left-most occurrence,
so if you added something at the end it will disappear and if you
added something at the beginning, the old one will disappear.
Thus the following works nicely in .zshenv:

  typeset -U path
  path=(~/bin ~/progs/bin $path)' .zshrc